# src/users/users.controller.ts
import { Controller, Get, Param, Query, Delete } from '@nestjs/common';
import { ApiOperation, ApiResponse, ApiParam, ApiQuery } from '@nestjs/swagger';
import { UsersService } from './users.service';
import { User } from './user.entity';

@Controller('users')
export class UsersController {
  constructor(private readonly usersService: UsersService) {}

  @Get()
  @ApiOperation({ summary: 'Get all users' })
  @ApiResponse({ status: 200, description: 'The found records', type: [User] })
  getUsers() {
    return this.usersService.findAll();
  }

  @Get(':id')
  @ApiOperation({ summary: 'Get user by ID' })
  @ApiParam({ name: 'id', type: 'number', description: 'User ID' })
  @ApiResponse({ status: 200, description: 'The found record', type: User })
  getById(@Param('id') id: number) {
    return this.usersService.findOne(id);
  }

  @Get('by-info/email')
  @ApiOperation({ summary: 'Get user by email' })
  @ApiQuery({ name: 'email', type: 'string', description: 'User email' })
  @ApiResponse({ status: 200, description: 'The found record', type: User })
  getByEmail(@Query('email') email: string) {
    return this.usersService.findOneByEmail(email);
  }

  @Delete(':id')
  @ApiOperation({ summary: 'Delete user by ID' })
  @ApiParam({ name: 'id', type: 'number', description: 'User ID' })
  @ApiResponse({ status: 200, description: 'Successfully deleted' })
  deleteUser(@Param('id') id: number) {
    return this.usersService.delete(id);
  }
}